Relocate JS importation
By placing <script> tags at the bottom of the <body> section, you can ensure that the rest of the page is rendered first, allowing the user to see and interact with the content as quickly as possible. Once the page has finished rendering, the scripts can be loaded and executed without affecting the user experience.